Sort events
If we run std::is_sorted it returns false sometimes, confirming the
event aren't always sorted.

std::stable_sort is used just in case a pair of 'B'egin and 'E'nd events
has the same 'ts' time. This guarantees the original order is maintained
in that case, although it could still be the case that the 'E'nd event
is before it's corresponding 'B'egin event. Fortunately, this rarely
happens in practice. Maybe we could just use std::sort for this same
reason?!